split lib into multiple modules
